[FIX] account: avoid duplicating account groups
